RowOfButtons created, Button.js refactored
